Lo de clase

+++ AndroidEngine/src/main/java/es/ucm/gdv/androidengine/AndroidGraphics.java
@@ -102,10 +102,10 @@ public class AndroidGraphics extends RescaleGraphics {
     @Override
     protected void finalDrawImage(Image img, Rect dest, Rect src){
         android.graphics.Rect aSrc = new android.graphics.Rect(
-                dest.get_left(),dest.get_top(),dest.get_right(), dest.get_bottom());
+                (int)dest.get_left(),(int)dest.get_top(),(int)dest.get_right(), (int)dest.get_bottom());
 
         android.graphics.Rect aDest = new android.graphics.Rect(
-                src.get_left(), src.get_top(), src.get_right(), src.get_bottom());
+                (int)src.get_left(), (int)src.get_top(), (int)src.get_right(), (int)src.get_bottom());
 
         _canvas.drawBitmap(((AndroidImage)img).getBitmap(), aSrc, aDest, null);
     } //finalDrawImage
